--- src/Engine/PostProcessing/Filter.hpp
@@ -1,24 +1,0 @@
-#pragma once
-
-namespace Video {
-    class ShaderProgram;
-}
-
-/// Post-processing filter.
-class Filter {
-    public:
-        /// Create filter.
-        Filter() { }
-        
-        /// Free allocated resources.
-        virtual ~Filter() { }
-        
-        /// Get shader program.
-        /**
-         * @return Shader program
-         */
-        virtual Video::ShaderProgram* GetShaderProgram() const = 0;
-        
-        /// Set uniforms.
-        virtual void SetUniforms() = 0;
-};
