+++ src/Engine/PostProcessing/GammaCorrectionFilter.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Filter.hpp"
+#include <Video/PostProcessing/Filter.hpp>
 #include <GL/glew.h>
 
 namespace Video {
@@ -8,7 +8,7 @@ namespace Video {
 }
 
 /// Gamma correction filter.
-class GammaCorrectionFilter : public Filter {
+class GammaCorrectionFilter : public Video::Filter {
     public:
         /// Create new color filter.
         GammaCorrectionFilter();
